+++ commerce/interaction-api/src/main/java/ru/yandex/practicum/commerce/dto/order/OrderDto.java
@@ -5,6 +5,7 @@ import lombok.AccessLevel;
 import lombok.Data;
 import lombok.experimental.FieldDefaults;
 
+import java.math.BigDecimal;
 import java.util.Map;
 import java.util.UUID;
 
@@ -34,9 +35,9 @@ public class OrderDto {
 
     boolean fragile; // признак хрупкости заказа
 
-    float totalPrice; // общая стоимость заказа
+    BigDecimal totalPrice; // общая стоимость заказа
 
-    float deliveryPrice; // стоимость доставки
+    BigDecimal deliveryPrice; // стоимость доставки
 
-    float productPrice; // стоимость товаров в заказе
+    BigDecimal productPrice; // стоимость товаров в заказе
 }
